requirements.txt
RPi.GPIO
smtplib
email
sqlite3
